Update mysql.sql
fix

+++ src/db/mysql.sql
@@ -3,8 +3,8 @@ CREATE DATABASE message_app;
 USE message_app;
 
 CREATE TABLE messages (
-	id bigint AUTO_INCREMENT NOT NULL,
+    id int AUTO_INCREMENT NOT NULL,
     content text NOT NULL,
-    createdUtc int NOT NULL,
+    createdUtc bigint NOT NULL,
     primary key (id)
 );
